requirements.txt
# Core AI / ML
torch==2.7.1
torchvision==0.22.1
transformers==4.53.2
accelerate==1.8.1
timm==1.0.17
einops==0.8.1
tokenizers==0.21.2
huggingface-hub==0.33.4
sentencepiece==0.2.0
omegaconf==2.3.0
safetensors==0.5.3

# OpenAI
openai==1.96.1

# numpy
numpy>=1.24,<2
scipy==1.15.3

# Vision
opencv-python  # do NOT pin, let pip pick version compatible with numpy<2
pillow==11.3.0
scikit-image==0.25.2
imageio==2.37.0

# Data / utils
matplotlib==3.10.3
pandas==2.3.1
tqdm==4.67.1
requests==2.32.4
fsspec==2025.7.0

# ROS2 Python nodes
rclpy==3.3.16
cv-bridge==3.2.1
message-filters==4.3.7

# Build / ROS code generation helpers
empy==3.3.4
catkin_pkg==1.0.0
setuptools==65.5.1
lark-parser==0.12.0
Cython==3.1.2
colcon-common-extensions

# Optional: YOLO
ultralytics==8.3.167
